creating inputTrip model form
Currently having issue with user_id column. Form does not save as is. User inputs should be saved to database, and admin can access these models. This worked before I implemented to user attribute for the inputTrip model

note: current edit_trip.html is a rough template. This will more than likely have to be another ModelForm, so that it can be saved in the database.