Added PackingList to Edit Trips Page
I made it so the packinglist is now directly on the edit trips page and you can add/delete/edit items directly from there. The AI should be connectded too so I think the packing-list functionality is completely done. The way it looks is kinda wierd because of the ircle in the background. I didn't know what was causing it